+++ lib/src/common/constants.dart
@@ -1,3 +1,4 @@
-const controlPointSize = 32.0;
+const controlPointSize = 62.0;
 const irregularCropItemLimit = 3;
 const rectangleCropItemLimit = 100.0;
+const cropLimit = 3.0;