+++ lib/src/widgets/crop/crop_area_clipper.dart
@@ -1,14 +1,14 @@
 import 'package:flutter/material.dart';
 
-import '../../common/picture_path_item.dart';
+import 'crop_area_item.dart';
 
 /// [CropAreaClipper] is used to clip the crop and guideline area by converting coordinate information to a [Path].
-/// [item] is a [PicturePathItem] that contains the coordinate information.
+/// [item] is a [CropAreaItem] that contains the coordinate information.
 /// [radius] determines the radius of the clip area.
 class CropAreaClipper extends CustomClipper<Path> {
   CropAreaClipper(this.item, {this.radius = 0.0});
 
-  final PicturePathItem item;
+  final CropAreaItem item;
   final double radius;
 
   @override